+++ hgicommon/tests/test_helpers.py
@@ -1,6 +1,6 @@
 import unittest
 
-from hgicommon.helpers import create_random_string
+from hgicommon.helpers import create_random_string, extract_version_number
 
 _PREFIX = "123"
 _POSTFIX = "456"
@@ -29,5 +29,25 @@ class TestCreateRandomString(unittest.TestCase):
         self.assertTrue(string.endswith(_POSTFIX))
 
 
-if __name__ == '__main__':
+class TestExtractVersionNumber(unittest.TestCase):
+    """
+    Tests for `extract_version_number`.
+    """
+    def test_when_no_version_number(self):
+        self.assertRaises(ValueError, extract_version_number, "no_version_number")
+
+    def test_when_major_version_number(self):
+        self.assertEqual("1", extract_version_number("test1version"))
+
+    def test_when_major_minor_version_number(self):
+        self.assertEqual("1.2", extract_version_number("test1_2version"))
+
+    def test_when_major_minor_patch_version_number(self):
+        self.assertEqual("1.2.3", extract_version_number("test1_2_3version"))
+
+    def test_when_multiple_version_numbers(self):
+        self.assertEqual("1", extract_version_number("test1or2"))
+
+
+if __name__ == "__main__":
     unittest.main()